requirements.txt
apache-airflow[gcp]
apache-beam[gcp]
pytest
pandas
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
dvc[gdrive]
